requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
PyJWT>=2.8.0
cryptography>=42.0.0
httpx>=0.26.0
aiofiles>=23.2.1

# Database
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Celery & Redis
celery>=5.3.6
redis>=5.0.1
flower>=2.0.1

# Additional utilities
python-json-logger>=2.0.7
tenacity>=8.2.3
